refactor(cli): extract reusable cli package from cmd/tachograph
Per reusable SDK CLI PRD, extract command definitions into a `cli/`
package with `NewCommand(opts ...Option)` so tachograph can be mounted
in the `way` CLI under `way connect tachograph`.

- New `cli/` module with separate go.mod (OSS 3-module pattern)
- `cmd/tachograph/main.go` reduced to thin entry point with fang wiring
- No functional changes to CLI behavior
