+++ cli/cli.go
@@ -1,0 +1,6 @@
+package cli
+
+// Option configures the CLI command tree.
+type Option func(*config)
+
+type config struct{}
